+++ Runtime/Int/Int.UniTask/Int.UniTaskFluxExtension.cs
@@ -28,16 +28,16 @@ namespace Kingdox.Flux
 #region UniTask
     public static partial class FluxExtension //Action<UniTask>
     {
-        public static void Subscribe(this byte key, Action<UniTask> action, bool condition) => Core.Flux<UniTask>.SubscribeActionParam(key, action, condition);
-        public static void Subscribe(this byte key, Func<UniTask> action, bool condition) => Core.Flux<UniTask>.SubscribeFunc(key, action, condition);
-        public static void Subscribe<T>(this byte key, Func<UniTask<T>> action, bool condition) => Core.Flux<UniTask<T>>.SubscribeFunc(key, action, condition);
-        public static void Subscribe<T>(this byte key, Func<T, UniTask> action, bool condition) => Core.Flux<T,UniTask>.SubscribeFuncParam(key, action, condition);
-        public static void Subscribe<T,T2>(this byte key, Func<T, UniTask<T2>> action, bool condition) => Core.Flux<T,UniTask<T2>>.SubscribeFuncParam(key, action, condition);
+        public static void Subscribe(this int key, Action<UniTask> action, bool condition) => Core.Flux<UniTask>.SubscribeActionParam(key, action, condition);
+        public static void Subscribe(this int key, Func<UniTask> action, bool condition) => Core.Flux<UniTask>.SubscribeFunc(key, action, condition);
+        public static void Subscribe<T>(this int key, Func<UniTask<T>> action, bool condition) => Core.Flux<UniTask<T>>.SubscribeFunc(key, action, condition);
+        public static void Subscribe<T>(this int key, Func<T, UniTask> action, bool condition) => Core.Flux<T,UniTask>.SubscribeFuncParam(key, action, condition);
+        public static void Subscribe<T,T2>(this int key, Func<T, UniTask<T2>> action, bool condition) => Core.Flux<T,UniTask<T2>>.SubscribeFuncParam(key, action, condition);
 
-        public static UniTask @UniTask(this byte key) => Core.Flux<UniTask>.TriggerFunc(key);
-        public static UniTask<T> @UniTask<T>(this byte key) => Core.Flux<UniTask<T>>.TriggerFunc(key);
-        public static UniTask @UniTask<T>(this byte key, T @param) => Core.Flux<T, UniTask>.TriggerFuncParam(key, @param);
-        public static UniTask<T2> @UniTask<T, T2>(this byte key, T @param) => Core.Flux<T, UniTask<T2>>.TriggerFuncParam(key, @param);
+        public static UniTask @UniTask(this int key) => Core.Flux<UniTask>.TriggerFunc(key);
+        public static UniTask<T> @UniTask<T>(this int key) => Core.Flux<UniTask<T>>.TriggerFunc(key);
+        public static UniTask @UniTask<T>(this int key, T @param) => Core.Flux<T, UniTask>.TriggerFuncParam(key, @param);
+        public static UniTask<T2> @UniTask<T, T2>(this int key, T @param) => Core.Flux<T, UniTask<T2>>.TriggerFuncParam(key, @param);
     }
 #endregion
 }
